Move sources.py out of mkwutil
